# Alternate nodeSet data structure