## Description

Uses an alternate data structure for the node set. For the operation of getOrInsert where the node doesn't exist (which is by far the most common case), we are replacing 9 map lookups and 1 map insert with 2 map lookups and 2 map inserts.

## Check List

Have you:

- Added unit tests? Relies on existing.

- Add cmprefimpl tests? (if appropriate?) Relies on existing.

## Related Issue

N/A

## Benchmark Results

Relevant tests:

```
COMPARISON
name                    old time/op    new time/op    delta
Intersection/n=10-4       61.7µs ± 8%    59.3µs ± 9%   -3.88%  (p=0.029 n=13+15)
Intersection/n=100-4       336µs ± 8%     311µs ±13%   -7.43%  (p=0.001 n=15+14)
Intersection/n=1000-4     3.07ms ± 4%    2.75ms ± 9%  -10.39%  (p=0.000 n=13+13)
Intersection/n=10000-4    33.5ms ±14%    28.9ms ±19%  -13.67%  (p=0.000 n=14+13)

name                    old alloc/op   new alloc/op   delta
Intersection/n=10-4       27.9kB ± 0%    29.2kB ± 0%   +4.52%  (p=0.000 n=15+15)
Intersection/n=100-4       144kB ± 0%     155kB ± 0%   +7.49%  (p=0.000 n=15+15)
Intersection/n=1000-4     1.72MB ± 0%    1.88MB ± 0%   +9.08%  (p=0.000 n=15+15)
Intersection/n=10000-4    18.1MB ± 0%    19.3MB ± 0%   +6.60%  (p=0.000 n=15+15)

name                    old allocs/op  new allocs/op  delta
Intersection/n=10-4          282 ± 0%       284 ± 0%   +0.71%  (p=0.000 n=15+15)
Intersection/n=100-4         446 ± 0%       457 ± 0%   +2.47%  (p=0.000 n=12+14)
Intersection/n=1000-4      2.06k ± 0%     2.07k ± 0%   +0.40%  (p=0.000 n=14+15)
Intersection/n=10000-4     19.1k ± 0%     19.1k ± 0%   +0.10%  (p=0.001 n=15+15)
```